LayoutTests/js/regexp-unicode-expected.txt

Test for unicode regular expression processing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "a".match(/a/u)[0].length is 1
PASS "a".match(/A/ui)[0].length is 1
PASS "a".match(/a/u)[0].length is 1
PASS "a".match(/A/iu)[0].length is 1
PASS "È’".match(/È’/u)[0].length is 1
PASS "È’".match(/È’/u)[0].length is 1
PASS "áˆ´".match(/áˆ´/u)[0].length is 1
PASS "áˆ´".match(/áˆ´/u)[0].length is 1
PASS "âª¼".match(/âª¼/u)[0].length is 1
PASS "ã¿­".match(/ã¿­/u)[0].length is 1
PASS "ğ’…".match(/ğ’…/u)[0].length is 2
PASS "ğ’…".match(/ğ’…/u)[0].length is 2
PASS "ğŒ†".match(/ğŒ†/u)[0].length is 2
PASS /ğ‘/u.test("ğ‘") is true
PASS /ğ‘/u.test("ğ‘") is true
PASS "ğŒ†".match(/ğŒ†/u)[0].length is 2
PASS /(ğ€€|ğ€|ğ©)/u.test("ğ€") is true
PASS "ğ„£".match(/a|ğ„£|b/u)[0].length is 2
PASS "b".match(/a|ğ„£|b/u)[0].length is 1
PASS /(?:a|ğ„£|b)x/u.test("ğ„£") is false
PASS /(?:a|ğ„£|b)x/u.test("ğ„£x") is true
PASS /(?:a|ğ„£|b)x/u.test("b") is false
PASS /(?:a|ğ„£|b)x/u.test("bx") is true
PASS "ağ„£x".match(/ağ„£b|ağ„£x/u)[0].length is 4
PASS /(ğ€€|ğ€|ğ©)x/ui.test("ğ€x") is true
PASS /(ğ€€|ğ€|ğ©)x/ui.test("ğ©x") is true
PASS /(ğ€€|ğ€|ğ©)x/ui.test("ğx") is true
PASS /(ğ€€|ğ€|ğ©)x/ui.test("ğ¨x") is true
PASS "ğ©".match(/a|ğ|b/iu)[0].length is 2
PASS "B".match(/a|ğ„£|b/iu)[0].length is 1
PASS /(?:A|ğ„£|b)x/iu.test("ğ„£") is false
PASS /(?:A|ğ„£|b)x/iu.test("ğ„£x") is true
PASS /(?:A|ğ„£|b)x/iu.test("b") is false
PASS /(?:A|ğ„£|b)x/iu.test("bx") is true
PASS "ağ„£X".match(/ağ„£b|ağ„£x/iu)[0].length is 4
PASS "Å¤x".match(/Å¥x/iu)[0].length is 2
PASS "ğŒ†".match(/^.$/u)[0].length is 2
PASS "It is 78Â°".match(/.*/u)[0].length is 9
PASS stringWithDanglingFirstSurrogate.match(/.*/u)[0].length is 3
PASS stringWithDanglingSecondSurrogate.match(/.*/u)[0].length is 3
PASS "ğŒ†".match(/[ğŒ†a]/)[0].length is 1
PASS "ğŒ†".match(/[ağŒ†]/u)[0].length is 2
PASS "ğŒ†".match(/[ğŒ†a]/u)[0].length is 2
PASS "ğŒ†".match(/[a-ğŒ†]/)[0].length is 1
PASS "ğŒ†".match(/[a-ğŒ†]/u)[0].length is 2
PASS "X".match(/[ -ğ‘]/u)[0].length is 1
PASS "á€€".match(/[ -ğ‘]/u)[0].length is 1
PASS "ğ§".match(/[ -ğ‘]/u)[0].length is 2
PASS re1.test("Z") is false
PASS re1.test("á€€") is false
PASS re1.test("ğ€") is false
PASS re2.test("A") is true
PASS re2.test("ï¿¿") is false
PASS re2.test("ğ’…") is true
PASS "ğŒ‘ğŒ‘ğŒ‘".match(/ğŒ‘*a|ğŒ‘*./u)[0] is "ğŒ‘ğŒ‘ğŒ‘"
PASS "ağŒ‘ğŒ‘".match(/ağŒ‘*?$/u)[0] is "ağŒ‘ğŒ‘"
PASS "ağŒ‘ğŒ‘ğŒ‘c".match(/ağŒ‘*cd|ağŒ‘*c/u)[0] is "ağŒ‘ğŒ‘ğŒ‘c"
PASS "ağŒ‘ğŒ‘ğŒ‘c".match(/ağŒ‘+cd|ağŒ‘+c/u)[0] is "ağŒ‘ğŒ‘ğŒ‘c"
PASS "ğŒ‘ğŒ‘ğŒ‘".match(/ğŒ‘+?a|ğŒ‘+?./u)[0] is "ğŒ‘ğŒ‘"
PASS "ğŒ‘ğŒ‘ğŒ‘".match(/ğŒ‘+?a|ğŒ‘+?$/u)[0] is "ğŒ‘ğŒ‘ğŒ‘"
PASS "ağŒ‘ğŒ‘ğŒ‘c".match(/ağŒ‘*?cd|ağŒ‘*?c/u)[0] is "ağŒ‘ğŒ‘ğŒ‘c"
PASS "ağŒ‘ğŒ‘ğŒ‘c".match(/ağŒ‘+?cd|ağŒ‘+?c/u)[0] is "ağŒ‘ğŒ‘ğŒ‘c"
PASS "ğŒ‘ğŒ‘ğŒ‘".match(/ğŒ‘+?a|ğŒ‘+?./iu)[0] is "ğŒ‘ğŒ‘"
PASS "ğªğªğŒ‘".match(/ğ‚*ğˆ€|ğ‚*ğŒ‘/iu)[0] is "ğªğªğŒ‘"
PASS "ğªğªğŒ‘".match(/ğ‚+ğˆ€|ğ‚+ğŒ‘/iu)[0] is "ğªğªğŒ‘"
PASS "ğªğªğŒ‘".match(/ğ‚*?ğˆ€|ğ‚*?ğŒ‘/iu)[0] is "ğªğªğŒ‘"
PASS "ğªğªğŒ‘".match(/ğ‚+?ğˆ€|ğ‚+?ğŒ‘/iu)[0] is "ğªğªğŒ‘"
PASS "abğŒ‘cğ¨".match(/abc|abğŒ‘cd|abğŒ‘cğ¨d|abğŒ‘cğ¨/u)[0] is "abğŒ‘cğ¨"
PASS "abğ¨cğ¨".match(/abc|abğ€cd|abğ€cğ¨d|abğ€cğ¨/iu)[0] is "abğ¨cğ¨"
PASS /abc|abğ€cd|abğ€cğ¨d|abğ€cğ¨/iu.test("qwerty123") is false
PASS "ağ¨ğ¨ğ¨c".match(/ac|ağ€*cd|ağ€+cd|ağ€+c/iu)[0] is "ağ¨ğ¨ğ¨c"
PASS "abğ¨ğ¨ğ¨cğ¨".match(/abc|abğ€*cd|abğ€+cğ¨d|abğ€+cğ¨/iu)[0] is "abğ¨ğ¨ğ¨cğ¨"
PASS "abğ¨ğ¨ğ¨".match(/abc|abğ¨*./u)[0] is "abğ¨ğ¨ğ¨"
PASS "abğ¨ğ¨ğ¨".match(/abc|abğ€*./iu)[0] is "abğ¨ğ¨ğ¨"
PASS "ğ€".match(/a*/u)[0].length is 0
PASS "ğ€".match(/a*/ui)[0].length is 0
PASS "ğ€".match(/\d*/u)[0].length is 0
PASS "123ğ€".match(/\d*/u)[0] is "123"
PASS "12X3ğ€4".match(/\d{0,1}/ug) is ["1", "2", "", "3", "", "4", ""]
PASS match3[0] is "ağğb"
PASS match3[1] is undefined.
PASS match3[2] is "ağğb"
PASS match4[0] is "ağ¸ğ¸b"
PASS match4[1] is undefined.
PASS match4[2] is "ğ¸ğ¸"
PASS match5[0] is "ağ’ğ’bğ’ğ’"
PASS match5[1] is undefined.
PASS match5[2] is "ğ’ğ’"
PASS match6[0] is "ağ’ğ’bğºğ’"
PASS match6[1] is undefined.
PASS match6[2] is "ğ’ğ’"
PASS /Å¿top/ui.test("stop") is true
PASS /stop/ui.test("Å¿top") is true
PASS /â„ªelvin/ui.test("kelvin") is true
PASS /KELVIN/ui.test("â„ªelvin") is true
PASS /\u{1}/.test("u") is true
PASS /\u{4}/.test("u") is false
PASS /\u{4}/.test("uuuu") is true
PASS "800-555-1212".match(/[0-9\-]*/u)[0].length is 12
PASS "this is ba test".match(/is b\cha test/u)[0].length is 11
PASS new RegExp("\\/", "u").source is "\\/"
PASS r = new RegExp("\\u{110000}", "u") threw exception SyntaxError: Invalid regular expression: invalid unicode {} escape.
PASS r = new RegExp("\\-", "u") threw exception SyntaxError: Invalid regular expression: invalid escaped character for unicode pattern.
PASS r = new RegExp("\\a", "u") threw exception SyntaxError: Invalid regular expression: invalid escaped character for unicode pattern.
PASS r = new RegExp("[\\a]", "u") threw exception SyntaxError: Invalid regular expression: invalid escaped character for unicode pattern.
PASS r = new RegExp("[\\b]", "u") threw exception SyntaxError: Invalid regular expression: invalid escaped character for unicode pattern.
PASS r = new RegExp("[\\B]", "u") threw exception SyntaxError: Invalid regular expression: invalid escaped character for unicode pattern.
PASS r = new RegExp("\\x", "u") threw exception SyntaxError: Invalid regular expression: invalid escaped character for unicode pattern.
PASS r = new RegExp("[\\x]", "u") threw exception SyntaxError: Invalid regular expression: invalid escaped character for unicode pattern.
PASS r = new RegExp("\\u", "u") threw exception SyntaxError: Invalid regular expression: invalid escaped character for unicode pattern.
PASS r = new RegExp("[\\u]", "u") threw exception SyntaxError: Invalid regular expression: invalid escaped character for unicode pattern.
PASS successfullyParsed is true

TEST COMPLETE
